Feature(Delete entity)
 Now, we can conveniently delete the entity, we just need to call actor.delete ().
Earlier, when deleting an actor, we did not know which components he left a
reference to himself, now there should be no memory leaks.
 When deleting, the actor places itself in the array for deletion
and the scene is safe, at the end of the frame will delete the actor
so as not to break the iteration over the systems.
 Also, the actor has a deleted flag, which means that this ather will be deleted
at the end of the current frame.